+++ cvmatrix/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "3.1.6"
+__version__ = "3.2.0"
 __all__ = ["CVMatrix", "Partitioner"]
 from .cvmatrix import CVMatrix
 from .partitioner import Partitioner